requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.8
anthropic==0.7.8
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
pandas==2.1.3
matplotlib==3.8.2
plotly==5.17.0
websockets==12.0
channels==4.0.0
channels-redis==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0
# Optional: Uncomment when needed
# django-celery-beat==2.5.0
# flower==2.0.1
